requirements.txt
numpy-1.20.1
transformers-4.4.2
pandas-1.2.3
scikit-learn-0.24.1
scipy-1.6.1
